# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Estate Hub — Real Estate ERP + CRM SaaS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /leads:
    get:
      operationId: getLeads
      tags: [leads]
      summary: List leads
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: assigned_to
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of leads
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeadListResponse"
    post:
      operationId: createLead
      tags: [leads]
      summary: Create a new lead
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadInput"
      responses:
        "201":
          description: Created lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
  /leads/{id}:
    get:
      operationId: getLead
      tags: [leads]
      summary: Get a single lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
    patch:
      operationId: updateLead
      tags: [leads]
      summary: Update a lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadUpdate"
      responses:
        "200":
          description: Updated lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
    delete:
      operationId: deleteLead
      tags: [leads]
      summary: Delete a lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /leads/{id}/assign:
    patch:
      operationId: assignLead
      tags: [leads]
      summary: Assign a lead to a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadAssign"
      responses:
        "200":
          description: Lead assigned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"

  # ─── CUSTOMERS ──────────────────────────────────────────────
...
  /leaves:
    get:
      operationId: getLeaveRequests
      tags: [leaves]
      summary: List leave requests
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: user_id
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of leave requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LeaveRequest"
    post:
      operationId: createLeaveRequest
      tags: [leaves]
      summary: Submit a leave request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeaveRequestInput"
      responses:
        "201":
          description: Created leave request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaveRequest"
  /leaves/{id}:
    get:
      operationId: getLeaveRequest
      tags: [leaves]
      summary: Get a single leave request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Leave request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaveRequest"
    patch:
      operationId: updateLeaveRequest
      tags: [leaves]
      summary: Approve or reject a leave request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeaveRequestUpdate"
      responses:
        "200":
          description: Updated leave request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaveRequest"

  # ─── DOCUMENTS ───────────────────────────────────────────────
...
components:
  schemas:
...
    Lead:
      type: object
      properties:
        id:
          type: integer
        company_id:
          type: integer
        name:
          type: string
        phone:
          type: string
        email:
          type: ["string", "null"]
        source:
          type: string
        status:
          type: string
        assigned_to:
          type: ["integer", "null"]
        assigned_user_name:
          type: ["string", "null"]
        follow_up_date:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        created_at:
          type: string
        updated_at:
          type: string
      required:
        - id
        - company_id
        - name
        - phone
        - source
        - status
        - created_at
        - updated_at
    LeadInput:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        source:
          type: string
        status:
          type: string
        assigned_to:
          type: ["integer", "null"]
        follow_up_date:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
      required:
        - name
        - phone
        - source
    LeadUpdate:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        source:
          type: string
        status:
          type: string
        assigned_to:
          type: ["integer", "null"]
        follow_up_date:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
    LeadAssign:
      type: object
      properties:
        assigned_to:
          type: integer
      required:
        - assigned_to
    LeadListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Lead"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - data
        - total
        - page
        - limit

    # ─── CUSTOMER SCHEMAS ────────────────────────────────────
...
    LeaveRequest:
      type: object
      properties:
        id:
          type: integer
        company_id:
          type: integer
        user_id:
          type: integer
        user_name:
          type: ["string", "null"]
        leave_type:
          type: string
        start_date:
          type: string
        end_date:
          type: string
        reason:
          type: ["string", "null"]
        status:
          type: string
        reviewed_by:
          type: ["integer", "null"]
        reviewed_at:
          type: ["string", "null"]
        created_at:
          type: string
        updated_at:
          type: string
      required:
        - id
        - company_id
        - user_id
        - leave_type
        - start_date
        - end_date
        - status
        - created_at
        - updated_at
    LeaveRequestInput:
      type: object
      properties:
        leave_type:
          type: string
        start_date:
          type: string
        end_date:
          type: string
        reason:
          type: string
      required:
        - leave_type
        - start_date
        - end_date
    LeaveRequestUpdate:
      type: object
      properties:
        status:
          type: string
        reason:
          type: string

    # ─── DOCUMENT SCHEMAS ────────────────────────────────────
...